scripts/utilities/requirements.txt
gspread==5.12.0
google-auth==2.24.0
requests==2.31.0
python-dotenv==1.0.0